requirements.txt
ImageHash==4.3.1
mss==7.0.1
numpy==1.23.5
Pillow==9.3.0
PyWavelets==1.4.1
scipy==1.9.3
